Complete split of LLM client out of POC service
